fix(diagnostic): fix backwards compatibility for goto_next and goto_prev
Summary:


8ba73f0e4cc6 deprecated `vim.diagnostic.goto_next` in favor of
`vim.diagnostic.jump`, and reimplements `goto_next` in terms of `jump`.  This
is fine.  However, when reimplementing `goto_next`, we forgot that the default
behavior of `goto_next` is to show a floating window with the error (the
`float` option).

This commit adds the `float` option to the `opts` table in `goto_next` and
`goto_prev`, and sets it to `true` by default.  This is the behavior that
`goto_next` and `goto_prev` had before the deprecation.  This commit also
updates the documentation to reflect this change.

Test Plan:

1. use neovim 0.10
2. run `:lua vim.diagnostic.goto_next()`
3. verify that the error is shown in a floating window
4. run `:lua vim.diagnostic.goto_next({float=false})`
5. verify that the error is not shown in a floating window

6. use previous commit to this commit
7. run `:lua vim.diagnostic.goto_next()`
8. verify that the error is NOT shown in a floating window

9. use this commit
10. run `:lua vim.diagnostic.goto_next()`
11. verify that the error is shown in a floating window

Fixes #29591